requirements.txt
selenium==4.5.0
python-dotenv==0.21.0
beautifulsoup4==4.12.2
requests==2.28.1
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.0
google-auth==2.15.0
google-api-python-client==2.70.0
google-api-core==2.11.0
googleapis-common-protos==1.57.0
openai==0.28.0
lxml